added new changes

+++ src/main/java/com/comment/commentservice/service/CommentService.java
@@ -1,16 +1,16 @@
 package com.comment.commentservice.service;
 
+import com.comment.commentservice.Exception.CommentNotFoundException;
+import com.comment.commentservice.constfiles.ConstFile;
 import com.comment.commentservice.feign.FeignLike;
 import com.comment.commentservice.feign.FeignUser;
 import com.comment.commentservice.model.CommentDto;
 import com.comment.commentservice.model.CommentModel;
 import com.comment.commentservice.repo.CommentRepository;
-import com.customer.Exception.CommentNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -28,26 +28,45 @@ public class CommentService {
 
 
     public int commentCount(String postId) {
-        int count = this.commentRepository.findByPostID(postId).size();
-       return  count;
+        List<CommentModel> allData = commentRepository.findAll();
+        int count = 0;
+        for (CommentModel comment : allData) {
+            if (comment.getPostID().equals(postId)) {
+                count++;
+            }
+        }
+        return count;
     }
 
 
-    public CommentModel updateComment(CommentModel commentModel, String postId, String commentId) {
+    public CommentDto updateComment(CommentModel commentModel, String postId, String commentId) {
         commentModel.setCommentID(commentId);
         commentModel.setUpdatedAt(LocalDateTime.now());
         commentModel.setCreatedAt(commentRepository.findById(commentId).get().getCreatedAt());
         commentModel.setPostID(postId);
-        return commentRepository.save(commentModel);
+       commentRepository.save(commentModel);
+        CommentDto commentDTO=new CommentDto(commentModel.getCommentID(),commentModel.getComment(),
+                feignUser.findByID(commentModel.getCommentedBy()),
+             feignLike.countLike(commentModel.getCommentID()),
+                commentModel.getCreatedAt(),commentModel.getUpdatedAt());
+        return commentDTO;
+
     }
 
 
     public String deleteByCommentId(String commentId) {
-        this.commentRepository.deleteById(commentId);
-        return "Delete CommentID " + commentId + " from DB";
+        if(commentRepository.findById(commentId).isPresent()){
+            this.commentRepository.deleteById(commentId);
+            return ConstFile.idDeleted;
+        }
+        else {
+            throw new CommentNotFoundException(ConstFile.idNot);
+        }
+
     }
 
     public List<CommentDto> allComments(String postId, Integer page, Integer pageSize) {
+
         if (page == null) {
             page = 1;
         }
@@ -60,10 +79,12 @@ public class CommentService {
         List<CommentDto> commentDTOS = new ArrayList<>();
 
         for(CommentModel commentModel:commentModels){
-            CommentDto commentDTO1=new CommentDto(commentModel.getCommentID(),
+            CommentDto commentDTO1=new CommentDto(commentModel.getCommentID(),   commentModel.getComment(),
                     feignUser.findByID(commentModel.getCommentedBy()),
-                    commentModel.getComment(),commentModel.getCreatedAt(),commentModel.getUpdatedAt(),
-                   feignLike.countLike(commentModel.getCommentID()));
+                    feignLike.countLike(commentModel.getCommentID()),
+                 commentModel.getCreatedAt(),commentModel.getUpdatedAt()
+                 );
+            commentDTOS.add(commentDTO1);
         }
         return  commentDTOS;
 
@@ -72,13 +93,15 @@ public class CommentService {
 
     public CommentDto saveComment(CommentModel commentModel, String postId) {
         commentModel.setPostID(postId);
+        commentModel.getComment();
         commentModel.setCreatedAt(LocalDateTime.now());
         commentModel.setUpdatedAt(LocalDateTime.now());
         this.commentRepository.save(commentModel);
-        CommentDto commentDTO=new CommentDto(commentModel.getCommentID(),
+        CommentDto commentDTO=new CommentDto(commentModel.getCommentID(),commentModel.getComment(),
                 feignUser.findByID(commentModel.getCommentedBy()),
-                commentModel.getComment(),commentModel.getCreatedAt(),commentModel.getUpdatedAt(),
-                feignLike.countLike(commentModel.getCommentID()));
+                feignLike.countLike(commentModel.getCommentID()),
+                commentModel.getCreatedAt(),commentModel.getUpdatedAt()
+               );
         return commentDTO;
     }
 
@@ -88,13 +111,14 @@ public class CommentService {
 try {
     CommentModel commentModel = commentRepository.findById(commentId).get();
 
-    CommentDto commentDTO = new CommentDto(commentModel.getCommentID(),
+    CommentDto commentDTO = new CommentDto(commentModel.getCommentID(),commentModel.getComment(),
             feignUser.findByID(commentModel.getCommentedBy()),
-            commentModel.getComment(), commentModel.getCreatedAt(), commentModel.getUpdatedAt(),
-            feignLike.countLike(commentModel.getCommentID()));
+            feignLike.countLike(commentModel.getCommentID()),
+            commentModel.getCreatedAt(), commentModel.getUpdatedAt()
+            );
     return commentDTO;
 }catch (Exception e){
-    throw  new CommentNotFoundException("comment id not found ");
+    throw  new CommentNotFoundException(ConstFile.idNot);
 }
 
     }
